test[vm_runtime] make LoadedModule::new infallible
This cannot fail. (The original plan was to have `LoadedModule::new` handle the case where the sender doesn't match the account `self` handle. But the correct place to handle that is in `VerifiedTransaction::new`, since this is a property of the transaction and not of the module.)

Further diffs will clean up surrounding code that is also infallible.