test[Consensus] Fix SMR tests flakiness due to Pacemaker jumpstart task
We saw a couple of flaky SMR tests: in all of them the problem was that the first round would time out (and the test checks would not expect that).
Now the fundamental solution is to modify the time service used in these tests (to control the timeouts).
Given the real clock we're using for timeouts we had a problem that the first timeout deadline is set before the runtime spawns the first "NewRoundEvent" task in the local pacemaker. Hence, if the initialization of SMR would take more time we'd hit the timeout and the tests would break.
This diff moves the timeout setup code to be executed from within the context of a task that is going to send the initial NewRoundEvent task.
Also increased the base timeout from 1sec to 3sec in the SMR tests.