test[Consensus] Fix process_votes_basic_test
Our process_votes_basic_test was a bit of a fake :)
It was looking for new round event for round 1, which is event that generated when pacemaker is initialized. (not related to vote, as the test name suggests).

Vote produces event for round 2, but we normally was just ignoring this event, since event from initialization would make test pass.

However, pacemaker was recently refactored to have part of initialization concurrent, as a result we sometimes had races, when event from vote process would appear before event from initialization, causing test to fail.

This fixes test by first waiting on initialization event before processing vote.

Bigger question is that our pacemaker seem to be a bit racy, but we can think of solution separately.