testEd25519 batch signature verification.
Batch signature verification as described in the original EdDSA article by Bernstein et al. "High-speed high-security signatures". Current implementation works for signatures on the same message and it checks for malleability.

note: This is for current `signing.rs` and we'll apply to nextgen when we finish with our crypto create refactoring.