testFix wrong table used in gas_schedule
The gas cost for each instruction is stored in a `CostTable` which is split into a `compute_table` and `memory_table`. Together these two costs (compute and memory) are used to calculate a total cost for an instruction during execution. There are two helper functions that get the cost from each table; `memory_gas` and `comp_gas`.

While I using MIRAI to verifying the safety of the `unwrap()` calls in each function, I came across the following copy and paste bug: the `comp_gas` function was using the `memory_table` rather than the `compute_table`. As a result, the compute cost being returned for each instruction is just 1.

A consequence of this fix is that the default maximum gas bound had to be raised and several tests that had to be modified to increase account balances.