test[functional tests] allow any stage to be disabled
Previously we only had options to disable the verifier (`//! no-verify`) and the runtime (`//! no-execute`). This allows any stage to be disabled. Here's the new syntax:
```
//! no-run: compiler, verifier, runtime
```

This will be useful whenever we add features to the language, as it allows one to test only the parts that are implemented, and gradually enable the rest.

Note if a critical stage like the compiler is disabled, all stages depending on it will also be skipped.