test[Cluster test] Simple cluster test that reboots one validator
This diff introduces simple 'skeleton' for cluster tests and have one simple experiment with rebooting validator.
It does not monitor metrics yet, this is going to be next part.

It also does certain things in simplified way, for example instead of having SSH client library it just uses ssh CLI, which is obviously sub optimal.

Right now it just runs experiments for rebooting random validator and waits for it to come back.