test[general] setup_panic_handler ensures all thread panics lead to process exit
Currently Tokio's default behavior is to catch panics and ignore them.
This lead to an issue where we observed a panic! that did not crash a Tokio
thread and led to a misbehaving process that wouldn't die.  Since
setup_panic_handler now always overrides sets the custom panic hook, this
won't happen again.

Closes 298.