testMove coverage script to scripts, allow reporting coverage for different submodules
The coverage report generation script in language/vm/ has been moved to scripts/. It has also
been augmented with an argument that allows the user to specify which tests should be run.
For instance, that argument can be '.' to run all tests (from the root of the libra directory)
or 'language' to run all the tests in the language directory.