test[docker] Improve build cache usage
Summary: Run the same build command (building both e2e and client) in
all docker images. This allows the docker cache from after the build to
be reused for all images, because all the rules up to this point are the
same. This will significantly speed up building all three images.

Test Plan: Built all three images; rust build was not repeated for the
second and third.