test[terraform] add a variable for adding capabilities as linux parameter
To run some experiments, we may need to add additional capabilities to the
ecs task. For example, to inject network delay using the `netem` tool,
`net_admin` capability is needed.
This PR adds the parameter for adding capabilities, but the default value
is empty which doesn't affect production. In the experiments we run,
just set this parameter explicitly to add the special capabilities.