test[storage] use distinct clients with distinct channels
This is some gRPC hackery going on here.
In this use cases, gRPC is used within the same process, so
optimizations that target the common case of multiple distinct
clients and small number of servers do not apply in our use case.
To break the aggressive connection pinning to completion queue
we add a distinct user agent, so connections are not grouped
together.
This works, it's not pretty.