test[execution][vm] Introduce Script and Module transactions.
We want to move away from Program transactions where execution and
publishing are in a single transaction.
There are several problems and complications with that model.
The plan is to have 2 kind of transactions:
- Script which are execution only transactions (CompiledScript)
- Module which are publishing only transactions (CompiledModule)

This diff defines the 2 type of transactions without removing Program yet.
This diff should be backward compatible and it should change nothing in the way
transactions are submitted and executed.
We are only defining the transaction type (Script/Module) and enabling
ser/deser tests for them. However the transactions are not supported yet by the VM
and they report an error if they are submitted to AC or the Executor.

The sequence of diffs and PRs is going to be the following:
- this PR introduces the transaction types
- next PR will enable the transaction in the VM
- then another PR we will move tests to use the new model
- finally we will have a PR that disables Program (that will be the breaking change)