test[storage][jellyfish][3/3] Jellyfish Sparse Merkle Tree implementation
This is the main implementation and tests of JellyfishMerkleTree. Which
is essentially is a sparse merkle tree as our current implementation but
with a few major optimizations.
1. We use NodeKey(version | nibble path) as key to uniquely identify a
node instead of hash. The prefix - version - will help reducing
compaction overhead since each commit of nodes will not overlap with any
preceding commit thanks to monotonically increasing version.
2. We remove the Extension node to simplify code. It can be proved that
the number of extension nodes will decrease as time goes and tree gets
denser, so we choose to get rid of it to make the code easier to
understand and maintain.
3. Merge blob node with leaf node for now for simplicity.