test[Consensus] Process QCs not only from proposals
Summary:
A validator can receive QCs via multiple ways: proposals, timeout certificates, QCs that are aggregated by the validator itself.
We used to send these QCs to SafetyRules and process the corresponding commits only if the QC is received via proposal.
However, once we introduce the NIL block voting we want to be able to commit blocks via the QCs that are generated during the timeouts: in this change we're adding QC processing to some several other places.

Tests: Extended the NIL block voting test to verify that a commit happens upon timeout.
Task: ref #371