test[storage] Pruner optimiaztion: seek to first unpruned record faster
With records being deleted on the smaller version side of the retire record log, seeking to
the first not-yet-deleted one gets slower and slower. While we can have more sophisticated impllementation for this FIFO type data structure, remembering the current version being processed between pruning sessions is a straightforward optimization.