test[language] Update gas meter to take into account that equality is now deep
Previously we relied on the builtin Rust equality operator for our
equality. However, that is no longer the case, and the equality checking
is now deep. Because of this, the gas cost of equality and inequality is
dependent upon the maximal size of the two values being compared. This
commit updates the gas metering for these instruction to reflect these
new cost characteristics.