test[storage] `TransactionStore::lookup_transaction_by_account()`
Built an index table to support looking up transaction versions by account_address and sequence_number. So that to serve the corresponding public API we don't need to binary search on the full state history, which in turn makes possible the pruning of state history on any node that serves read APIs.