test[types] proptest refactor to ease generating logical `TransactionToCommit` batches
Introducing the `AccountInfoUniverse` and `xyzGen` types.
The universe keeps track of a collection of accounts along with sequences numbers and event handles (which in turn has sequence numbers tracks).
The `xyzGen` types resolves "local" randomness uppon creation and when they `materialize()`, get/update "global" info in the arena.

For example, a `RawTransactionGen` has determined Payload upon creation, while gets and increases the account sequence number in the arena when it `materialize()`.

This way:
1. We don't need things like `renumber_events()` that's hacky and confusing.
2. Similarly, easy to support more "real" transaction sequence numbers in transaction batches.