test[cluster_test] Use intermediate tag for cluster_test validators
Previously ECS in cluster_test was configured to use `nightly` tag.
This introduced problem, because we don't control when nightly tag changes. If it changes during reboot experiment, validator come back with newer code, that might be incompatible with existing cluster.

This PR changed cluster_test to use separate `cluster_test` image tag. When `nightly` tag changes, we first re-tag this image with `cluster_test` tag, and then update cluster. This makes sure we are in control of a image version that cluster test currently running