test[language] bump sequence number for genesis transaction
This commit adjusts the genesis transaction code to bump the sequence number for the Association account, which is the sender of the genesis transaction. If we don't do this and a subsequent transaction (e.g., minting) is sent from the Assocation account, a problem arises: both the genesis transaction and the subsequent transaction have sequence number 0. This change ensures that every account satisfies the expected sequence number invariants.