test[swarm][config] Enable config generation for full node swarm
A full node swarm can be connected to a valdiator which acts as the
upstream for state sync protocol.

To start a full node swarm, users need to run:
`./target/debug/libra_swarm -n 4 -s -l -r "full_node" -u <UPSTREAM_PATH>`
where UPSTREAM_PATH is the config directory for the node which is its
upstream.

As part of full node config generation, we also modify the upstream
node's configuration to add a full node network, with the members of the
full node swarm as its network peers. As a result, the upstream
validator node needs to be restarted to kick-off the state sync to the
full nodes, since we don't have a way to dynamically load configuration
into a running process.

Closes: #1051
Approved by: phoenix-antigravity