test[canonical_serialzier] Implement a more generic api for serializing maps.
Currently when serializing a BTreeMap, we are just sorting on the serialized result of the key. This means that we really don't rely on the fact that the key is sorted. As a result, we can relax the api a bit so that it will just take a iterator of key-value tuples and serialize them

Closes: #1109
Approved by: davidiw