test[Consensus] Insert vote can generate TimeoutCertificate
Summary:
We extend the VoteReceptionResult enum to include TimeoutCertificate, which is returned by the block tree if there are 2f+1 votes for the same round.
Note that QC is higher priority than the TC, hence in case there are 2f+1 votes for the same ledger info digest, QC is returned.
The returned TimeoutCertificate value is not used by the event processor yet, to be done in the followup commits.

Testing:
updated the unit tests checking vote aggregation logic

Closes: #1132
Approved by: dmitri-perelman