test[crypto] un-mask GENESIS_BLOCK_ID
This removes the masking flag on GENESIS_BLOCK_ID. It was introduced
in #1477, but put on a `lazy_static` macro, which expands into several
interreltated declarations. As the masking flag only applied to the
lexically first of them, masking it, its dependent would fail to find
that first declaration.

The result would be a
```
error[E0573]: expected type, found static `GENESIS_BLOCK_ID`
``
on macro expansion.

Closes: #1498
Approved by: zekun000