test[StateSynchronizer] Revert the flow that would allow sync to an arbitrary target.
Summary:
We were trying to come up with a simplified solution that would allow both the validators and the full nodes
operate in a similar fashion by trying to state sync to the highest possible ledger info.

Unfortunately this flow significantly complicates the logic in consensus due to the new reconfiguration design
because it opens a possibility for an EventProcessor to accidentally state sync to another epoch. The basic
paradigm is for the EventProcessor to operate in one epoch only, hence complicating that logic is not worth the
effort.

In addition to that we realized that the new flow did not really solve the liveness attack of non-reachable
state sync. Instead, we are planning to update storage to support idempotent commits, which would mean that
in future state sync should be able to just "give up" and return without changing the visible storage state.

Testing: existing unit test coverage

Closes: #1600
Approved by: zekun000