test[language] add get and set for Vector
I was reading some code that used the struct MyStruct<T> pattern and was temporarily confused--why do we allow this? I remembered that there was a good answer and that a motivating use-case is in Vector: this lets us have a single Vector implementation with a strict API for resource types and a more permissive API for unrestricted types.

However, the current Vector API doesn't have any procedures that allow a more permissive API for unrestricted types. This PR fixes that by adding two such procedures: get and set. These examples demonstrate why the pattern makes sense + makes vector nicer to use for unrestricted types.

Closes: #1601
Approved by: davidiw