test[x] skip whitespace lints for .exp files
In some future patch a number of .exp files are going to be added to the
repository. These files are used for testing the move source language
and include the expected output from the move compiler. Since these
files are the project of a third-party library its difficult to
completely control their format and ensure that there are no whitespace
violations. Due to this, lets just skip whitespace lints for all .exp
files.

Closes: #1634
Approved by: tnowacki