test[cluster-test] Test if file exists before log rotate
When cluster test fails to setup cluster(for example, failed to cp genesis.blob), then there is no log file on host - previous was log rotate, and new one was not created.

In this case attempt to log rotate it on next run produces a lot of noise.

This checks if file exists before attempting to log rotate it

Closes: #1692
Approved by: ankushagarwal