test[config] a meager attempt at a test
This only verifies that the default config can be parsed but doesn't
ensure equality. For equality tests, we're going to need to do a lot
more work such as defining what is in scope for equality and what is
not. For example, this test does not verify that the data_dir_path is
not accidentally typed out as data_dir and so this is part 1 of a 2 part
testing infra for configs.

Closes: #1722
Approved by: mimoo