test[fuzzing] adding fuzzing coverage code
This PR adds a test that will run the fuzzers on every corpus file we give it.
This is to obtain fuzzer coverage statistics.

It expects to be called with a CORPUS_PATH environment variable like this:

```
CORPUS_PATH=fuzz/corpus cargo tarpaulin -p libra-fuzzer --ignored -- coverage
```

If the corpus path is `fuzz/corpus` (the default) then CORPUS_PATH can be ommitted.

Closes: #1805
Approved by: sausagee