test[config] use lcs for genesis.blob and load genesis in config
per title, eliminated the weird proto wrapper and just use lcs directly.

In additon, one of the recent gripes was that we didn't check genesis
earlier in the config loading. This move genesis directly into the
config allowing that check to happen as well as not requiring us to save
the genesis config to disk during testing.

Closes: #1815
Approved by: mimoo