test[config] simplify validator network generation
Like the previous commit, this leverages the various struct methods to
generate a validator network. As a result, we able to begin deleting a
lot of helper code. However, that helper code has a few more
dependencies that need to be cleaned up before we can completely remove
it.

Closes: #1830
Approved by: bmwill