test[storage] grpc backup protocol
Updated with working full client / server test for the
gRPC service.

The biggest change apart from adding the service itself was
moving the jellyfish merkle tree iterator out of the state
store and into libradb proper.  This was done so it could
be owned there in an Arc, allowing it to have a proper
lifetime for use in the futures-01 Stream.

Closes: #1901
Approved by: zamsden