test[State sync] Convert chunk request / response from protobuf to rust types.
Summary:
Libra as a project is gradually moving away from using protobuf for field descriptions: we utilize LCS for serialization and do not plan
versioning at this level anyway.
In this change we introduce the Rust structs GetChunkRequest and GetChunkResponse. Working with Rust structs is more straightforward in the code,
and as we're going to add more complexity with the enums for Waypoints it doesn't make sense to support it in Protobuf.
We're still using the Protobufs as the containers for the messages (to distinguish between different message types), similarly to the way it's done in Consensus.

Testing: existing test coverage

Issues: ref #1384

Closes: #1930
Approved by: dmitri-perelman