test[language] Add new parse_comma_list function to the parser
This new function consolidates all the separate loops for parsing
comma-separated lists into one place. This helps make the code
more consistent and readable, which will be especially valuable
when this code is enhanced to provide better error messages.
(We may or may not do that for the IR compiler, but it is
important for Move source compiler.)

Besides adopting the new function is places where it is a drop-in
replacement, this makes a slight change to parsing ArgDecls and
FieldDecls, which both had optional commas after them. With this
change, all but the last comma in a list of ArgDecls or FieldDecls
are now mandatory. I updated one place in libra_account where a
comma was missing after a FieldDecl.

Closes: #1931
Approved by: vgao1996