test[Storage] Last version of an epoch should belong to the same epoch
Summary:
The last version of an epoch used to be mapped to the next epoch because of the "empty block" concerns
(in case the very first ledger info in the next epoch is an empty block the same version would belong to both epochs).

However, we do not allow empty blocks anymore, so this concern is not valid now.
Being able to map the last version of an epoch to the same epoch is important for Waypoints and client processing in general:
* in case of waypoints a client should be able to retrieve the LedgerInfo the waypoint certifies
* specifically we want a client to be able to retrieve the initial validator set given the genesis waypoint

In this commit we change the indexing of versions to the epochs: the index keeps the last version of the epoch.
One consequence is that a client with version `0` is always receiving the initial validator set back.

Testing: updated the existing tests.

Issues: ref #1384

Closes: #1948
Approved by: wqfish