test[vm] Fix soundness bug in VMStr::new
The previous API was unsound. The following program demonstrates a heap
Use After Free:

    use vm::vm_string::VMStr;

    fn main() {
        let s = "...........".to_owned();
        let v = VMStr::new(s);
        println!("{}", v);
    }

This program instantiates VMStr::new with 'a = 'static. The type String
satisfies impl AsRef<str>+'static so the borrow checker is fine with it.
Note that String is 'static because it does not borrow any data so it
lives as long as the caller wants, not restricted by any lifetime
shorter than 'static.

Ownership of the owned string `s` is moved into VMStr::new and dropped
at the end of new. The reference returned by VMStr::new then points to a
freed part of the heap.

On my machine this minimal repro prints a variety of exciting content,
such as:

    h`É
       KV

but could just as easily segfault or leak secrets in real usage.

Closes: #1949
Approved by: bmwill