test[storage] make epoch proof consistent with get_epoch
We change the API of get_epoch_change_ledger_infos to accept a end_epoch
instead of a version.

This change helps us delegate the need for last epoch to the caller and
solves both client and validator use case as well as preparing for
resolving TooManyRequested.

Closes: #1961
Approved by: wqfish