test[Storage] Kill mock genesis
One confusing thing in tests right now is that the genesis ledger info
comes from the mock, and the rest of them come from test input (normally
something like `arb_blocks_to_commit`). However, when retriving the
ledger infos from DB, the DB is supposed to return all of them. That's
why [this test](https://github.com/libra/libra/blob/647f02a77af56eedd889fdbd3b0028acb14907d7/storage/libradb/src/libradb_test.rs#L19)
needs to do the `remove(0)` stuff.

Another benefit of this is that it unifies the genesis ledger info and
others to treat them the same way in storage. One minor thing to note is
that real genesis ledger info is signed by empty validator set, but in
tests it might be signed normally. This shouldn't really matter.

Overall this just simplifies things a bit and makes the code easier to
understand, note that before this change there were a couple of
`version+1`s in the tests which is not intuitive.

Closes: #1991
Approved by: lightmark