test[language] remove use of get_gas_remaining from epilogue
This opcode is only used in the transaction epilogue. We don't want it in the language because it makes the gas semantics observable from the concrete semantics (e.g., can branch on the value of get_gas_remaining).

This PR replaces the opcode with an explicit gas_units_remaining parameter to the epilogue + makes the VM fail if it encounters the get_gas_remaining opcode. It does not yet delete the opcode from the file format--we are waiting to make all of the breaking changes to the format at once.

Closes: #1997
Approved by: vgao1996