test[language][move] Improve parser error messages for comma-separated lists
Replace some "unexpected token" errors with messages that are specific
to comma-separated lists. This changes the parse_comma_list function
to also consume the starting and ending tokens, so that it can report
better messages about them. For type arguments, where the starting "<"
character is attached to the type name token, this requires
identifying the source location for that "<" character and also a
separate parse_comma_list_after_start entry point. To simplify the code
a bit, this also adds a new match_token function that combines the
operation of checking for a token and then advancing past it.

Closes: #2067
Approved by: tnowacki