test[State sync] Integration tests for waypoint synchronization
Summary:
Some refactoring of the integration tests to allow for:
* chains of peers, in which a peer serves as an upstream to the next one.
* not starting all the peers at once

As a result, we can now run tests that verify chains of state sync.
We can also build a test, in which peer 0 runs to epoch 10, a waypoint is taken based on its LedgerInfo of epoch 7, peer 1 starts with this waypoint and finishes its initialization, peer 2 sync ups to peer 1.

Testing: all these tests

Issues: ref #1384

Closes: #2028
Approved by: zekun000