testAdd flag to ValidatorChangeEventWithProof to indicate whether it is complete
Changed the server to not return any response items in
update_to_latest_ledger request if the response does not have all
necessary LedgerInfos for the client to update to latest one.

Also realized that now GetEpochChangeLedgerInfosResponse is identical to
ValidatorChangeEventWithProof, so remove the struct entirely.

This change finishes the required server side changes for
https://github.com/libra/libra/issues/1923

Closes: #2037
Approved by: zekun000