test[language] Fix distribution of transaction fees to handle dust
If the amount in the fee pot was less than the number of validators,
then this would cause an error when `deposit`ing the coin. We now check
this property before entering the distribution loop, and if the amount
to be paid is zero, we skip the loop.

Closes: #2039
Approved by: sblackshear