test[consensus] Add safety rules benchmark
The intent of this benchmark is to help evaluate design choices for the
TCB prior to full e2e evaluation. It generates. updates, and votes on
blocks -- performing the most critical and frequent actions for the TCB.
It is entirely possible that the benchmark will be moderately consumed
by just generating data structures but it should still be sufficient to
identify regressions / improvements to the LSR code.

Closes: #2043
Approved by: davidiw