testcluster-test: temporary fix
During the migration process to using tonic for the AC grpc endpoint I
was under the impression that all uses of the ac client from within
cluster-test were done in sync contexts. This turned out to be incorrect
and resulted in trying to build two tokio runtimes (one within another)
which results in the code panicing.

Longer-term we'll need to do some refactoring so that cluster-test can
use the new clients, but in the short term lets just reintroduce a
grpcio version of the AC client that cluster-test can use.

Closes: #2077
Approved by: ankushagarwal