testAdd type checking to spec translator.
In order to deal with things like field selection and correct treatment of references, we need type information for the translated SpecExp's. This change adds type checking to the translator and extends the test suite.

The type checking also has the added benefit that we report context errors in the spec language directly at the source instead of via the generated boogie. Right now, the tool bails out on the first type error, but the code is setup so systematic collection of errors with source level backlink can be added in subsequent PRs.

Closes: #2083
Approved by: sblackshear