test[consensus][easy] fix the epoch race
There's some reace between epoch manager and the networking that results
in starting the same epoch EventProcessor multiple times.

Because we may receive multiple EpochChange messages and before we start
the new epoch they'll all pass the verification and stay in the queue.

This change adds a guard to the avoid this.

Closes: #2096
Approved by: dmitri-perelman