testUnified environment and utilities for move prover.
This is a large refactoring which has been in the make (on and off) for a while.

- Introduces a new object model for representing the environment for the translator (in `env.rs`). The object model allows to navigate a set of loaded modules and retrieve all information needed for compilation to boogie. Some of that info comes from the bytecode, other from the AST. While at this point the environment is part of the move prover, it is designed to be independent, so we can move it somewhere else at a later point if desired.
- Rewrites the bytecode translator (now in file `bytecode_translator.rs`) to make use of the new environment. Size is reduced by ~40%. Similar rewrites `spec_translator.rs`.
- Solves the problem of needing 'ghost declarations' because the environment uses a `GlobalType` instead of `SignatureToken` which is valid accross modules.

All existing tests pass, and the golden files have been updated and manually verified for absence of regressions. Functionality wise, this is mostly a NO-OP, beside of that we

- now uses the actual parameter names found in the move AST.
- for now removed debugging code in the bytecode translator. This wasn't connected yet, and will be brought back once we work on this feature.

This change does not make use of features of the new environment yet (e.g. collecting of error messages instead of bailing out after each error) which is done in a future PR.

Closes: #2303
Approved by: sblackshear