test[Storage] Add backup and restore command line tools
This implements a few basic things. The backup tool will ask the server
for the latest version, then using the version to initiate the
backup_account_state request. Then it splits the stream into small
chunks according to some chunk size and creates a file for each chunk.
The restore tool will take the files as input, then construct an
iterator of account chunks and proofs and use them to restore libradb.

Closes: #2070
Approved by: lightmark