test[consensus] pass in author directly
Currently we have a special config for consensus that is derived from
the underlying config + the author field. I suspect there are some
legacy reasons for this, but in its current form, eliminating author
from the ChainedBftSMRConfig allows removal of ChainedBftSMRConfig in
favor of ConsensusConfig.

This commit adds in the new path for author.

Closes: #2151
Approved by: dmitiri-perelman