test[consensus] Remove signer from chained_bft_smr_test
This means we need to introduce configs into chained_bft_smr_test as
that is how we configure safety rules. We also need a means to clone the
test configs across executions, so the test_config clone code needed to
be updated to support the correct testing behavior.

As a result of this refactor, we were able to also simplify the way in
which we were generating the nodes.

Closes: #2173
Approved by: zekun000