test[language][move] Remove the NameBeginArgsValue token
There is an ambiguity in the grammar for Move where a "<" character
following a name in an expression term may be either a binary
operator or the start of a list of type arguments. The Move parser
does not currently have the capability to use name lookup to resolve
this ambiguity based on the type of the name. Instead we have been
relying on the presence of whitespace before the "<" character when
it is used as a binary operator. This change continues that approach
but does so by adding a targeted check for whitespace in the one
place where the ambiguity exists, instead of the previous approach
where the "<" character was "glued" to the name as a special
NameBeginArgsValue token.

Closes: #2193
Approved by: tnowacki