testdev-setup: remove unneeded dependencies
Now that we have finished migrating off of grpcio to using tonic we no
longer need to have developers install a number of extra dependencies
that were necessary to build grpcio (since it is a wrapper around a C
lib).

Remove these extra dependencies from the dev-setup script as well as all
of our Dockerfile.

Closes: #2258
Approved by: sausagee