test[language] add Offer module
Say you have a resource of type R stored somewhere under account address A and you want to give it to address B. For example: say I want to create a recoverable account by delegating the KeyRotationCapability of address A to address

B. How can you do this? Unless B has a published resource exposing an API that accepts an R, there is no way to make this exchange.

The Offer module introduced in this PR solves this problem. Now, I can send a transaction from A to create an Offer<KeyRotationCapability> under address A, and B can claim it using Offer.claim<KeyRotationCapability>(a). In the near future, this will be useful for transferring KeyRotationCapability and WithdrawalCapability's between accounts during the setup of custody systems.

Closes: #2281
Approved by: kchalkias