test[secure] Add disk (persistent) backend for secure storage
This is largely a copy of the persistent data on disk used in
SafetyRules and is intended to replace it. The biggest difference is the
switch from using two files for handling persistent writes and instead
writing to a temp and then moving it to the long term location. This
simplifies the system a bit as now only one file will ever be valid.

Closes: #2313
Approved by: zekun000