test[CI] MIRAI is not a check anymore, but a repo dispatch
Github actions are limited for pull requests coming from fork repositories.
This is because when our github action runs, the GITHUB_TOKEN provisioned by Github is the one from the forked repository.
This makes it impossible for the github action to comment on the PR if it finds any issues with MIRAI.

In order to fix this, we will:

- run this action only locally, if triggered (via the `repository_dispatch` hook)
- trigger this remotely via a private bot that detects new PR

Closes: #2357
Approved by: davidiw