test[safety-rules] Implement a reliable safety rules server
If a read / write fail, log the output and try to read again. The
NetworkServer library will internally attempt to establish a connection
with a new downstream peer.

Note: our current logging facilities employ substantial amounts of
unsafe code and cannot be used as is in the TCB.

Closes: #2459
Approved by: zekun000