test[move-prover] Introduce invariants in the mvir parser.
This adds ast and parser for invariants in mvir structs and resources, with a syntax as such:

```
resource T {
  i: u64,
  invariant i > 0
}
```

There is also a (more experimental) version of an invariant over a state delta:

```
resource T {
  i: u64,
  invariant {update} old(i) < i
}
```

The identifier in the braces -- called a modifier -- is free text while we figure out what we may need here.

We also add "synthetic" (ghost) variables for use in specifications:

```
synthetic sum_of_T_i: u128;
resource T {
  i: u64,
  invariant {update} sum_of_T_i == old(sum_of_T_i) - old(i) + i,
  invariant {unpack} sum_of_T_i == old(sum_of_T_i) - old(i),
  invariant {pack}   sum_of_T_i == old(sum_of_T_i) + i,
}
```

These features in the MVIR parser allows us to experiment with the design which will eventually be  in the source language once we've moved the spec language to there. They should be considered experimental for now.

Closes: #2475
Approved by: DavidLDill