test[language][move] Add parser support for bitwise shift operators and type casts.
These changes are intended to accompany the addition of u8 and u128 types
to Move. This adds both left and right shift operators and a type cast syntax.
The ast nodes for these operations are not yet available, so they are temporarily
stubbed out to use existing nodes.

Closes: #2533
Approved by: tnowacki