test[stdlib] Directly port over the current IR stdlib into Move
[language] Swap over to Move for non-testing functions

This commit swaps the stdlib and transaction scripts from the IR to
Move.

[language] Swap over to Move for non-testing functions

This commit swaps the stdlib and transaction scripts from the IR to
Move.

[language] Add staging to the stdlib

This PR stages the stdlib. The use of a "fresh" stdlib and scripts can
be controlled via the "MOVE_NO_STAGED_STDLIB" env variable.

Closes: #2705
Approved by: bob-wilson