test[move source language] Change "move" and "copy" to be unary expressions
The previous Move grammar allowed "move" and "copy" expressions to be
"Terms" used with field references, e.g., "move x.field". This was
misleading and could result in confusing error messages from the compiler.
It does not matter so much with "copy" but for "move" you have to move
the entire value before you access the fields. This changes them to be
unary expressions so that they need to be wrapped in parenthesis before
accessing fields, e.g., "(move x).field".

Closes: #2643
Approved by: tnowacki