test[scripts] maintain the cargo_update_outdated script
The awk command in this script retrieves source-compatible updates in the output of a cargo outdated run and applies them to the Cargo files in the workspace.
cargo-outdated added a column in its output, changing the position of the output portion we want to capture. This adapts to it apporpriately.

Closes: #2589
Approved by: metajack