test[move-prover] Specification language type checker and translator.
This implements type checking of specifications as they come out of the move compilers expansion phase. The expansion AST is translated into a new AST and integrated into the prover's environment abstraction.

The checker and AST is in a new crate `move-prover/spec-lang`. The `env.rs` from bytecode-to-boogie has been copied and adapted for the new design. Test infra for golden (baseline) tests has also been copied and put into a new create `move-prover/test-utils`.

The type checker implements type inference and is fairly complete (no known feature gaps). It uses a unified approach to binary and unary operators and specification function resolution. Tests are in `spec-lang/tests/sources`.

This PR doesn't finalize the ingestion of the checked and translated ast into the prover environment, but uses the environment only as far as needed for the type checking problem. A subsequent PR will deal with this problem, as well as adapting the new environment and ast for translation to Boogie.

Closes: #2766
Approved by: davidldill