test[KS][GHA] add land-blocking test kill switch
The land-blocking test workflow depends on the cluster test in this repo
and the infra backing in another.  Before we turn it on, we need a KS to
bypass the workflow for ninja land or breaking the inherent dep loop.
Condsider this scenario where the cluster test itself breaks, no PRs can
land as they all fail the test, including the PR that fixes the breakage
in cluster test.

The cluster test has kill swich mechanism built in. That depends on
actions behind the scene in AWS. Here we perform the bypass at the
source.

Note that bors will be listening for the check status from the
land-blocking flow. We will let the workflow run but terminate early
with success exit status.

Closes: #2655
Approved by: andll