test[terraform] Add variable to set a permissions boundary
This variable can be set to a policy ARN which will be used as the
permissions boundary on roles created by Terraform. This can be used
to prevent these roles from being used to escalate AWS access.

Test Plan: Created workspace with and without setting permissions
boundary.

Closes: #2824
Approved by: opsguy