test[network] Refactor how dial requests are routed
Today, we need to construct ProtocolId::from_static("DialPeer") to send
a dial request to PeerManager. This was a hack to migrate to QoS when
sending requests to PeerManager (by having fairness over (PeerId,
ProtocolId)). However, as we migrate to new network wire protocol which
has a static set of Protocols listed in the spec, it's a good time to
remove the hack and create a new channel to send ConnectionRequests to
PeerManager.

Closes: #2928
Approved by: phlip9