test[language] fix flaky rotation capability test
This test uses an address as an authentication key, which should fail with abort code 12 after the switch to short addresses in a270ff0b3ece7cd610f8d0e61475d439f529eafc. However, it appears that functional testing framework only detects the failure intermittently for some reason. This is why the test was flaky. That issue is being investigated/fixed separately.

This PR fixes the failing test by saving the old authentication key and restoring it instead of using the address as the auth key. I also needed to add a procedure for extracting the auth key to LibraAccount.

Closes: #2921
Approved by: runtian-zhou