test[consensus] converge on Signature -> Ed25519Signature
This removes the cryptoproxy component for signature. The challenge was
whether or not to use type aliasing. In favor of making the code as
readable and simple as possible, I elected to eliminate any sort of
aliasing. Tools like sed or search and replace can help swap to new
types.

Closes: #2937
Approved by: zekun000