test[move-prover] Added core contracts as specification tests.
baseline_tests.rs::verify_or_update_baseline had a bug where it failed tests because of whitespace. The problem turned out to be that the .exp file (reference output) was stripped of trailing whitespace to make github happy, but the output file it was compared with was not stripped.  I wrapped a call to clean_for_baseline around the "text" arg to diff.

Copied core contracts from language/stdlib to move-prover/tests/sources/stdlib.  I kept the modules and transaction_scripts directories.

Only the modules/libra_coin.move module is specified. I copied the
others so we can add specifications to them.

libra_coin.move has updated versions of the specs for the .mvir version. I.e., aborts_if and ensures for individual functions.  There are interesting global properties to be specified, but I haven't tried doing that yet.

I also added dependency comments at the beginnings of the .move files so the prover can find them during "cargo test".

Three files have no Move specifications but still can't make it through the Move prover: libra_system.move, libra_block.move, and transaction_fee.move. Since the message is about a precondition not holding, the problem must be related to the "requires" that is automatically added to say that the transaction sender exists.  I have not investigated this problem yet.  The tests all pass because the ".exp" files have the erroneous messages in them.

Closes: #2996
Approved by: wrwg