test[config] Add in Config the ability to specify a namespace (prefix) for Vault
This handles the issue of pushing the namespace through configuration.
Along the way, I noticed a couple other issues and am too lazy to make
additional commits:
* Renamed Box<dyn Storage> methods to new_storage from new_#name_storage
(pretty redundant to have the name of the storage in the function name)
* Revealed the reset method with the intent of having a testing feature
on Cargo in the future -- this enabled safety rules test for Vault to
pass.
* Cleaned up initialization of safety rules in config builder since it
was a subset of the validator

Closes: #3031
Approved by: mimoo