test[move core modules] add ApprovedPayment module
New core module that allows a payee to approve payments with a cryptographic signature. The basic flow is:
(1) Payer sends payment metadata to the payee
(2) Payee signs metadata and sends 64 byte signature back to the payer
(3) Payer sends an approved payment to the payee via transaction invoking containing the payment metadata and a signature. The transaction will abort if the signature check fails.

Note: approved payment's are an accounting convenience/a courtesy mechansim for the payee, *not* a hurdle that must be cleared for all payments to the payee. In addition, approved payments do not have replay protection (that is, the approval signature can be replayed).

Closes: #3188
Approved by: tzakian