test[network] Fix how disconnects are handled
A notion of `connection_id` is introduced. ConnectionId is a unique
local identifier assigned to each connection by the ConnectionHandler.
The ConnectionId is part of the ConnectionMetadata struct that is passed
from PeerManager to Peer when we start the downstream modules to handle
IO over the connection. The metadata is passed back from Peer to
PeerManager on disconnect. This allows PeerManager to uniquely determine
which connection was lost. This is unlike earlier, when we had a (buggy)
heurisitic to determine this using connection origin.

Design decisions:
* We increment the LIBRA_NETWORK_PEERS counter on any new connection.
  Similarly, we decrement the counter on any lost connection. This means
  that LIBRA_NETWORK_PEERS may temporarily double-count a peer if there
  are 2 connections with it. However, in steady state the counter should
  reflect the number of unique peers we are connected to.
* Duplicate LostPeer notifications: It is possible that we send many
  duplicate LostPeer notifications for a peer to the upstream client.
  However, this does not affect correctness, since the notification is
  only sent when there are no active connections.

More details can be seen in: https://github.com/libra/libra/issues/3128

Closes: #3200
Approved by: davidiw