test[move-prover] Fixing a bug in update invariants in presence of branching.
Until now, we had wrongly assumed that a mut ref can only have exactly one before-borrow snapshot associated. However, due to branching this is not the case. Consider e.g.:

```
let r: &mut R;
if (b) r = &mut t1 else r = &mut t2;
```

Here `r` can represent either `&mut t1` or `&mut t2`, depending on the value of `b`.

In this PR, we extend the tracking of before-update index from a single value per reference to a set of indices. In the example above, the reference `r` will have assigned both the before-borrow indices of `&mut t1` and `&mut t2`. In order to decide which invariant has to be enforced when `r` dies, we use a new boolean variable `$before_borrow_N_used` which is managed at runtime and set to `true` once the associated borrow instruction is executed.

This also fixes a bunch of related issues, like incorrectly asserting invariants for borrows which are never executed because of branching. There are likely more issues with this if it comes to loops, and we eventually will need to move the current ad-hoc analysis we do on the bytecode to the DFA framework and peform it in a better principled way.

Closes: #3280
Approved by: shazqadeer