test[secure-storage] Rename some methods to better match their intents
We don't get private keys, we export them -- this matches the Vault API,
which is more explicit than the get API -- as this will be a different
permission.

We don't generate private keys, we create them -- this matches the KV API, generate didn't match any existing APIs.

Closes: #3276
Approved by: joshlind