test[move-prover] Stackless bytecode refactoring.
This refactors stackless bytecode to make it better suitable for a program transformation approach of compilation to boogie.

- The number of opcodes has been drastically reduced by merging similar codes into one. For example, `Add(r, x, y)` becomes `Binary(Add, r, x, y)`. While a large flat enum of opcodes is good for the VM, it is not good for transformations which in many cases can deal with a class of basic ops generically (i.e. do something for all `Binary(op, r, x, y)`).
- Any dependencies from file_format in the bytecode have been removed. For example, instead of `SignatureToken` the bytecode uses now our own `Type`. Also, type instantiations are not longer represented by a signature pool index but directly.

On the way some incorrect treatment of generic type instantiations (likely introduced by file_format refactoring) has been fixed. There seem to be not much test coverage on prover side for this, so it didn't cause a problem until now.

Unfortunately, a set of unit tests for the stackless bytecode generator had to be removed. Those tests where still using old IR and have no way to depend on `spec-lang` which we now need. However, test coverage should be maintained because the translator is now exhaustively used by the prover itself.

Closes: #3331
Approved by: shazqadeer