test[language] loader
Define a Loader as the struct/API that loads Modules and Scripts and
holds the cache of all loaded entities.
A Resolver is the struct/API used by the Interpreter (runtime execution).
Define also a bunch of runtime structures that are used instead of binary
format data to favor execution.

In the process removes the arena, self referential structures and those
"unsafe" crate (unsafe as in using Rust unsafe features)

Closes: #3385
Approved by: vgao1996