test[network] Add ErrorCode::NotSupported to wire protocol
ErrorCode::ParsingError does not capture the case where we support a
protocol in the code, but don't want to expose it to a client. An
example might be the internal protocols like the one that will be created
for Validator -> SafetyRules communication - these protocols can be
correctly parse, but should not be exposed to external nodes.

Closes: #3436
Approved by: metajack