test[move-lang] Extending parser and expansion by generic spec vars.
This PR adds the syntax `name<t1, .., tn>` resp. `mod::name<t1,..,tn>` to the asts and syntax of move lang. Move prover needs to have this expression forms to deal with global specifications of generic resources (like the new generic coin type).

As with other expression constructs only allowed in specs, we parse this form and then reject it in expansion if it doesn't appear in a spec block.

There is some change in error reporting implied by this change. Previously, if we had `name<t1,..,tn>`, an error would have been produced that either a `(..)` or a `{..}` needs to follow. Instead, now we accept this during parsing, but reject it in expansion. The existing test cases for this have been taken over, but now produce different baseline output. Because those tests where also the only case where we test the hint 'Perhaps you missed a space before `<`' a new test case was added to cover this error.

Closes: #3438
Approved by: tnowacki