test[cli] fix incorrect argument check for compile command
This is a follow up to #3425 where I changed the arguments for
the compile command. I missed that the minimum number of arguments
is now 3 instead of 4.

This also changes the help message to refer to "dependency_source_files"
with underscores instead of spaces to separate the words. That is more
consistent with the other messages.

Closes: #3516
Approved by: runtian-zhou