test[move-lang] Parser cleanup for GlobalCalls and ModuleAccess
Wolfgang's changes in #3438 laid the groundwork for some simplifications
around the way that we use the ModuleAccess nodes. Besides simplifying the
parser, I'm looking into adding support for name aliases and plan to move
toward using the ModuleAccess AST nodes to represent all the different kinds
of names. (I'd like to rename "ModuleAccess" but I'm leaving that for a
separate PR.)

- Extend the ModuleAccess AST node to also include global names.
- Distinguish Calls and GlobalCalls in the expansion phase, not the parser.
- Introduce a new "NameExp" non-terminal in the grammar (instead of the
  parse_pack_or_call_or_generic_name helper function) and refactor the
  code that uses it.

Closes: #3634
Approved by: tnowacki