testUse formally-verified crypto by default
The construction is now full Rust extracted directly from Coq at:
https://github.com/dalek-cryptography/curve25519-dalek/compare/master...calibra:fiat2
which brings in the formally verified code of:
https://github.com/mit-plv/fiat-crypto

This has been unit-tested continuously since last August, without a single break (see the present commit, which switches the place where we used to run those unit tests on formally verified crypto to testing the non-certified dalek backend, to ensure we can always come back to it).

Our construction is also running in fiat-crypto's continuous integration tests, which protects us from a regression on the Coq side.
https://github.com/mit-plv/fiat-crypto/blob/master/etc/ci/test-fiat-rust.sh

Since this is mature and well-understood, we can switch to using it by default.

Closes: #3548
Approved by: metajack