test[manager] Add the ability to generate a validator-config transaction
The intent here is to build a tool to generate genesis as well as add
validators after genesis. Hence, the tool will always build a
validator-config in the context of a transaction. There is still a lot
of work to do:

* Add access to the blockchain for actually constructing a transaction
that would allow adding a validator after genesis
* Waiting for updates to the new validator config
* Migration of Multiaddr to networkaddr

Testing now is quite simplistic, just verifying that the operation
doesn't fail. In the next series of commits, this will actually be
tested in depth by generating and verifying genesis.

Closes: #3615
Approved by: zekun000