test[move-prover] Refactor test suite driver and copy new lib into test tree.
This refactors our test directory and driver as follows:

- all top-level tests have been moved into `sources/functional`
- a copy of the new standard library has been placed in `sources/new_stdlib`
- tests don't need to longer declare their dependencies via the `// dep` comment. Instead, dependencies are automatically detected by using a search path.
- different test directories (like functional, stdlib, etc.) can use different flags and search paths. This is configured in `testsuite.rs`, and allows for example `new_stdlib` to find dependencies at a different place than `stdlib`.
- for `new_stdlib`, verification is off by default and need to be turned on using `pragma verify = true` (either in a module spec block for the whole module or per function spec block)
- a `.exp` file is not longer generated if the test passes without any errors. On UPBL=1, existing .exp files are instead removed if there are no errors. This prepares us to run directly against the standard library source of truth at a later point.

There is a nasty issue with native functions of the same name having different signatures in old and new library. To work around this, a hack was added to generate different boogie names so the old/new lib sources can be used by the same prelude. The native functions for the new library, as well as new vector function `index_of`, are not implemented yet and `assert false`.

Closes: #3609
Approved by: DavidLDill