test[management] secure_backend to dyn storage
This introduces a means to convert from the new command-line storage
parameter to an actual storage backend.

The format for the command-line is
"k0=v0;k1=v1;..."
where at least one of the k_i must be backend and the rest are options
for that backend.

The test vectors show what should be parseable and what is not. Some
advanced parameters, for example, the server address or token for Vault
will only be tested upon use.

Closes: #3640
Approved by: joshlind