test[language] Native Function crate break up
This is another step to rationalize native function for v1.
It's a lot of moving code more than real changes.
We break common native functions in their own crate.
The overall native function story is structured in the following way:
- move-vm/types/src/natives/function.rs contains all the import you need
to write a native functions. It's a simple module with no logic
- move-vm/runtime/src/native_functions.rs contains the linkage of the
runtime to native functions. That is where native functions are
declared and it's inside the runtime. That is a restriction for
now and one I am happy to push in our codebase.
- move-vm/natives is the crate that has most of the native functions.

I cleaned up code a bit in the process.
Next steps are in 2 directions:
1- review and possible changes to the arguments of a native function.
We may cobine things, simplifies API,...
2- better model for side effect, but we want to see how
the value/type/serializer story evolves first

So all things considered this is a good step in a more clear
encapsulation

Closes: #3702
Approved by: vgao1996