test[move-lang] Capture documentation comments, as well as support block comments.
This PR suggests to extend the move-lang compiler to be able to capture documentation comments and pass them to upstream tools.

This is achieved by extending the existing `strip_comments` function to return a map from `codespan::Span` to comment strings. This map is then passed into the lexer which allows to check and consolidate correct placement of doc comments, with minimal intrusion for the parser. The doc comments are not stored in the AST, but managed as a separate data structure.

The comment stripper has also been extended to support block comments in addition to line comments. Block comments may become important for larger documentation embedded in Move sources, and are also handy for development as they easily allow to comment out larger regions of code. Block comments can be nested.

In order to give reasonable error messages in the presence of comments (unbalanced block comments, displaced documentation comments), the source on which the move lang compiler reports errors does now include the comments. This let to some baseline file updates where comments where previously stripped in error messages.

Closes: #3714
Approved by: tnowacki