test[move-prover] Repro of the termination/performance problem.
This PR adds a testcase to `tests/sources/regression/performance_200511.move` which has been extracted from the issues we have with Libra account. The example has been attempted to minimize, but the behavior is so erractic that full minimization is difficult. Instructions are in the test source. This is a situation where an error is expected to be reported, but instead Z3 hangs. If the error is corrected, or the source changed in other ways, termination is fast. This applies both for use without and with array theory.

Closes: #3835
Approved by: junkil-park