test[fuzzing] adding storage structured fuzzer
This PR adds a new structured fuzzer for storage.
Meaning we can now fuzz functions that expect a struct as input (as opposed to a bytearray).

The way it works is that the fuzzer gives the function a random string of bytes,
and the function pretends that the random bytearray is the ouptut of an RNG it can use to construct the struct.
This is called a "passthrough" RNG and we use the proptest library to build such structs from random data.

The current behavior of a passthrough RNG (in the proptest library we depend on)
is to crash if the input is not large enough to build the struct.
This is not ideal, so to make it work the PR also patches proptest to make pass-through RNGs return 0s after they depleted the input given to them.

Closes: #3872
Approved by: msmouse