test[management] rewrite tests to use ondisk storage
This is an incremental step to allowing this framework to be used for
integration tests. When the original library was written, ondisk storage
couldn't be used due to lack of namespacing. Now it has namespacing and
can be used. This allows us to remove the ignore tag and also use these
frameworks for other types of tests, which is the motivation for
StorageHelper.

Closes: #3878
Approved by: joshlind