test[management] Ability to create waypoints and insert them into storage
This actually skips one step, which is Party A creates a waypoint and
maybe stores it somewhere.
Then Party B downloads waypoint and stores it in their secure store.

We don't be using that protocol for now, so this better maps to what
we're trying to accomplish.

Closes: #3922
Approved by: joshlind