test[x-lint] re-do default workspace members
Include the packages that are actually built by Cargo, not the ones ignored
by it in default builds.

In particular, I noticed that the `libra-proptest-helpers` crate was
included in default members even though it shouldn't have been.

This also moves the cache to right next to the PackageGraph -- this has clearer
semantics and makes some future uses more convenient. Unfortunately, this
requires a self-referential struct, which I use the `rental` crate for.

Closes: #3973
Approved by: bmwill