test[client] Detect and report failed transactions
The client code has not been checking the VM status to determine
if transactions executed successfully. This changes to check for
a status of EXECUTED and reports an error otherwise. This exposed
an issue where the external_transaction_signer test in smoke_test.rs
has been failing because it ran out of gas, so this also adjusts
the gas parameters for that test to make it pass.

https://github.com/libra/libra/issues/3663

Closes: #4017
Approved by: thefallentree