test[language] Make verifier aware of generic vs non generic instructions
This is a fix for issue #3893.
The verifier never checked for generic instructions correctly referring
to generic member and non generic ones referring to non generic members.
This fix it by adding that check in the type safety pass.
It is not clear that is the best pass to check for this property and
maybe a single pass for generic or something similar may be a better
choice. We will find out in time...

Closes: #4041
Approved by: tnowacki