test[move core modules] add_all_currencies parameter for parent/child vasp creation
One remaining inconvenience of parent and child VASP account creation is the need to send separate transactions to add new currencies to an account. This PR solves the problem by adding a add_all_currencies parameter to both the parent and child VASP creation scripts. When true, this initializes the freshly created account with all of the available currencies in the system.

VASPs that want to accept a subset of the currencies will still need to send an initial creation transaction followed by a few AddCurrency transactions, but that seems reasonable (and preferable to having a different script for every combination of currencies).

Closes: #4059
Approved by: n4ss