test[secure-storage] Implement an optional ability to import keys
May the crypto gods forgive me... but moving forward, we will probably
have cluster test and various test networks using vault yet test
networks depend on the ability to generate keys deterministically. So we
introduce a means to "import" keys into various secure storages. Because
this is for testing only, it is optional to implment and users should be
extremely careful using it.

Closes: #4119
Approved by: joshlind