test[language] Remove transaction size argument to MoveVM
This removes the size as an argument to the runtime.
It also removes the charging intrinsic gas from the runtime altogether.

I started with passing the size to the `CostStrategy` "constructor" and
have the `Interpreter` call `charge_intrinsic_gas()` (as one of the
possible solutions we talked about) but then realized it had the
unpleasant side effect to make "charging intrinsic" part of any single
call into the runtime, which seems hardly what we want.
The way it is is where I ended up after few iterations as it
was the cleanest I could think of.

@runtian-zhou, @tzakian please let me know your thoughts

Closes: #4149
Approved by: runtian-zhou