test[network] Remove unauthenticated from config
We require noise even if not fully authenticated to ensure that peers
have distinct ids. The work to enforce ids is still in progress, but
we're getting there.

As a result, the default config should have a dedicated network id even
if it is ephemeral. So now when we load a config, we will either find
one or generate one, eliminating the need for the None option for
IdentityKey, woot.

Closes: #4181
Approved by: zekun000