test[move core modules] replace Role resource with role_id field
The LibraAccount::Role<RoleData> resource is an unfortunate way to implement account roles. The problem is that we need to add wrapper functions inside LibraAccount to read or write any of the RoleData (see the ugly wrappers for VASP and DD functionality as an example).

This PR solves the problem by:
- Getting rid of LibraAccount::Role
- Adding an integer field role_id to LibraAccount that encodes a unique integer for each role
- Publishing the role-specific resources at the top-level rather than in RoleData. This uses the publishX(creator: &signer, created: &signer) pattern for each role-specific resource X.

The rest of the PR is just cleanup that this change enables: getting rid of all the wrapper functions, all the json-rpc and test views for the role resource, and so on.

Closes: #4406
Approved by: moeziniaf