test[move-prover] Let spec-lang frontend handle scripts which use a different name than `main`.
The frontend did still hardwire the function name in a script as `main`, though this restriction was removed a while ago. This PR fixes this.

This was discovered via a docgen run (but would have also triggered the MIRAI bot via general prover run), therefore the test added is in docgen.

Closes: #4419
Approved by: tzakian