test[stdlib] Updating LibraTimestamp and other changes.
This changes the definition of `is_genesis` to not depend on acquiring the CurrentTime resource. There are also multiple smaller
changes here and there (using CoreAddresses also in specs and others). Also changes the doc generator to distinguish
resources from structs, which leads to larger diff in generated docs.

Submitting this now to not accumulate to many merge conflicts.

Closes: #4722
Approved by: sblackshear